fix: resolve all 36 audit findings (P0, P1, P2)
P0 fixes:
- CORS preflight blocked by auth middleware (OPTIONS exemption)
- Dynamic Tailwind class construction in platform-breakdown
- Non-monetary metrics formatted as currency in Insights UI
- .dockerignore file for lean Docker builds

P1 fixes:
- MCP credentials warning when using mock fallback
- get_campaign_summary uses SQL aggregation (O(1) queries)
- Alembic migration system initialized with initial migration
- Rate limiting via slowapi on cost-incurring endpoints
- Docker container runs as non-root user
- Async DB operations offloaded to thread executor in mcp_sync
- React state-in-effect anti-pattern fixed across 3 pages
- Removed unused @radix-ui/react-select and @radix-ui/react-slot deps
- Sync paths documented (POST /api/campaigns/sync vs /api/mcp/sync)
- Test lifespan handler no longer seeds demo data
- Request body size limit middleware (1 MB cap)
- AI analyzer catches specific exceptions (OpenAIError, JSONDecodeError)
- Structured AI analysis columns (strengths, weaknesses, recommendations)
- Timeouts on OpenAI client and MCP tool calls
- Platform type validation in sync service

P2 fixes:
- MCP sync single-transaction pattern with per-campaign error handling
- campaign_name resolved via Pydantic model_validator (schema/insight.py)
- Pinned dependency versions (~= compatible release)
- CI workflow: removed redundant pip install ruff
- Budget optimizer confidence score no longer capped at 0.95
- mcp-servers pyproject.toml for proper package install
- Error boundary component wrapping app layout
- window.location.reload() replaced with load() re-fetch
- Seed data counts derived from actual created records
- image_url Pydantic validation (max_length=1000)
- Defensive MCP sync response structure parsing
- Consistent HTTP status codes (404/502 for MCP errors)
- Float → Numeric(12,4) for monetary fields across all models
- _normalise_status → _normalize_status (American spelling)
- Division-by-zero test for campaign summary
- Platform sync partial failure rollback handling
- Fragile test assertion tightened to exact campaign count

+++ backend/app/models/campaign.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import datetime
 from decimal import Decimal
 
-from sqlalchemy import DateTime, Float, Integer, String, func
+from sqlalchemy import DateTime, Float, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.models.base import Base
@@ -25,15 +25,15 @@ class CampaignModel(Base):
     status: Mapped[str] = mapped_column(
         String(50), default="active", comment="active, paused, completed, archived"
     )
-    daily_budget: Mapped[Decimal | None] = mapped_column(Float, nullable=True)
-    total_budget: Mapped[Decimal | None] = mapped_column(Float, nullable=True)
-    spent: Mapped[Decimal] = mapped_column(Float, default=0.0)
+    daily_budget: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    total_budget: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
+    spent: Mapped[Decimal] = mapped_column(Numeric(12, 4), default=0.0)
     impressions: Mapped[int] = mapped_column(Integer, default=0)
     clicks: Mapped[int] = mapped_column(Integer, default=0)
     conversions: Mapped[int] = mapped_column(Integer, default=0)
-    cpa: Mapped[Decimal | None] = mapped_column(Float, nullable=True)
+    cpa: Mapped[Decimal | None] = mapped_column(Numeric(12, 4), nullable=True)
     roas: Mapped[Decimal | None] = mapped_column(Float, nullable=True)
-    revenue: Mapped[Decimal] = mapped_column(Float, default=0.0)
+    revenue: Mapped[Decimal] = mapped_column(Numeric(12, 4), default=0.0)
     ctr: Mapped[Decimal | None] = mapped_column(Float, nullable=True)
     start_date: Mapped[datetime.datetime | None] = mapped_column(DateTime, nullable=True)
     end_date: Mapped[datetime.datetime | None] = mapped_column(DateTime, nullable=True)
